refactor: 소스 코드 패키지 이동
- 테스트 패키지에서 작성 중인 클래스들을 전부 소스 코드 패키지로 이동시켰습니다.
- 생성자에 롬복을 적용했습니다.
