+++ src/main/java/roomescape/theme/controller/ThemeController.java
@@ -1,5 +1,6 @@
 package roomescape.theme.controller;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -8,19 +9,16 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.theme.ThemeService;
 import roomescape.theme.dto.ThemeCreateRequest;
 import roomescape.theme.dto.ThemeResponse;
+import roomescape.theme.service.ThemeService;
 
+@RequiredArgsConstructor
 @RequestMapping("/themes")
 @RestController
 public class ThemeController {
 
     private final ThemeService themeService;
-
-    public ThemeController(ThemeService themeService) {
-        this.themeService = themeService;
-    }
 
     @PostMapping
     public ResponseEntity<ThemeResponse> create(
@@ -43,4 +41,3 @@ public class ThemeController {
         return ResponseEntity.noContent().build();
     }
 }
-
