+++ src/main/java/roomescape/theme/repository/JdbcThemeRepository.java
@@ -5,7 +5,6 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.theme.ThemeRepository;
 import roomescape.theme.domain.Theme;
 
 @Repository
