+++ src/test/java/roomescape/theme/repository/JdbcThemeRepositoryTest.java
@@ -9,7 +9,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.theme.ThemeRepository;
 import roomescape.theme.domain.Theme;
 
 @JdbcTest
