+++ src/test/java/roomescape/theme/service/ThemeServiceTest.java
@@ -9,8 +9,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
-import roomescape.theme.ThemeRepository;
-import roomescape.theme.ThemeService;
+import roomescape.theme.repository.ThemeRepository;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.dto.ThemeCreateRequest;
 import roomescape.theme.dto.ThemeResponse;
